Crate risc0_zkp

The RISC Zero Proving System

These are the core algorithms that prove and verify the execution of zk-STARK circuits.

Crate Feature Flags

The following crate feature flags are available.

Feature	Target(s)	Implies	Description

cuda		prove, std	Turns on CUDA GPU acceleration for the prover. Requires CUDA toolkit to be installed.

metal	macos	prove, std	Deprecated - Metal GPU acceleration for the prover is enabled by default on Apple Silicon.

prove	all except rv32im	std	Enables the prover, incompatible within the zkvm guest.

std	all		Support for the Rust stdlib.

Modules

adapter

Interface between the circuit and prover/verifier

core

Core module used to implement a zk-STARK prover and verifier.

field

Finite field types and operations

hal

layout

Utilities for accessing buffers of circuits symbolically.

taps

verify

Cryptographic algorithms for verifying a ZK proof of compute

Macros

digest

Macro for constructing a Digest from a hex string.

Constants

FRI_FOLD

FRI folding factor is 2 ^ FRI_FOLD_PO2

INV_RATE

Inverse of Reed-Solomon Expansion Rate

MAX_CYCLES

MAX_CYCLES_PO2

MIN_CYCLES

MIN_CYCLES_PO2

MIN_PO2

QUERIES

50 FRI queries is sufficient to achieve our security target of 97 bits (conjectured security)

ZK_CYCLES