risc0_zkp
Module adapter

Interface between the circuit and prover/verifier

Structs

CircuitStepContext

MixState

PolyExtStepDef

ProtocolInfo

A protocol info string for the proof system and circuits. Used to seed the Fiat-Shamir transcript and provide domain separation between different protocol and circuit versions.

Enums

PolyExtStep

Constants

PROOF_SYSTEM_INFO

Versioned info string for the proof system.

REGISTER_GROUP_ACCUM

REGISTER_GROUP_CODE

REGISTER_GROUP_DATA

Traits

CircuitCoreDef

traits implemented by generated rust code used in both prover and verifier

CircuitInfo

CircuitProveDef

traits implemented by generated rust code used in only the prover

CircuitStep

CircuitStepHandler

PolyExt

PolyFp

TapsProvider

Type Aliases

Arg

Var