risc0_zkp::core::hash
Module poseidon2

An implementation of Poseidon2 targeting the Baby Bear.

Re-exports

pub use self::consts::CELLS;

pub use self::consts::ROUNDS_HALF_FULL;

pub use self::consts::ROUNDS_PARTIAL;

pub use self::consts::ROUND_CONSTANTS;

Structs

Poseidon2HashSuite

A hash suite using Poseidon2 for both MT hashes and RNG

Poseidon2Rng

A random number generator driven by Poseidon2

Constants

CELLS_OUT

The size of the hash output in cells (~ 248 bits)

CELLS_RATE

The ‘rate’ of the sponge, i.e. how much we can safely add/remove per mixing.

M_INT_DIAG_HZN

Standardizing on the same coefficients as https://github.com/HorizenLabs/poseidon2.git (as used in its plain implementation of Poseidon2 for BabyBear)

Functions

poseidon2_mix

The raw sponge mixing function

unpadded_hash

Perform an unpadded hash of a vector of elements. Because this is unpadded collision resistance is only true for vectors of the same size. If the size is variable, this is subject to length extension attacks.