risc0_zkp::field::baby_bear
Struct ExtElem

pub struct ExtElem(/* private fields */);

Instances of ExtElem are elements of a finite field F_p^4. They are represented as elements of F_p[X] / (X^4 + 11). This large finite field (about 2^128 elements) is used when the security of operations depends on the size of the field. The field extension ExtElem has Elem as a subfield, so operations on elements of each are compatible. The irreducible polynomial x^4 + 11 was chosen because 11 is the simplest choice of BETA for x^4 + BETA that makes this polynomial irreducible.

Implementations

impl ExtElem

pub const fn new(x0: Elem, x1: Elem, x2: Elem, x3: Elem) -> ExtElem

Explicitly construct an ExtElem from parts.

pub fn from_fp(x: Elem) -> ExtElem

Create an ExtElem from an Elem.

pub const fn from_u32(x0: u32) -> ExtElem

Create an ExtElem from a raw integer.

pub fn const_part(self) -> Elem

Return the base field term of an Elem.

pub fn elems(&self) -> &[Elem]

Return Elem as a vector of base field values.