risc0_zkp::field::goldilocks
Struct Elem

pub struct Elem(/* private fields */);

The Goldilocks class is an element of the finite field F_p, where P is the prime number 2^64 - 2^32 + 1. Here we implement integer arithmetic modulo P for both Goldilocks and for a field extension of Goldilocks.

The Fp datatype is the core type of all of the operations done within the zero knowledge proofs, and is the smallest ‘addressable’ datatype, and the base type of which all composite types are built. In many ways, one can imagine it as the word size of a strange architecture, and its operations as wrapping operations which respect word size P.

The Fp class wraps all standard arithmetic operations to make finite field elements appear like ordinary numbers (which, for the most part, they are).

Implementations

impl Elem

pub const fn new(x: u64) -> Elem

Create a new Goldilocks field Elem from a raw integer.